refactor: move partial package data type to types
