+++ src/hydrate/values/types.ts
@@ -4,3 +4,11 @@ export interface HydrationInputValues extends InputValues {
 	author: string;
 	email: string;
 }
+
+export interface PartialPackageData {
+	author?: { email: string; name: string } | string;
+	description?: string;
+	email?: string;
+	name?: string;
+	repository?: string;
+}
